установил вебпакер
