+++ .gitignore
@@ -27,13 +27,13 @@
 /storage/*
 !/storage/.keep
 
-#/public/assets
+/public/assets
 .byebug_history
 
 # Ignore master key for decrypting credentials and more.
 /config/master.key
 
-#/public/packs
+/public/packs
 /public/packs-test
 /node_modules
 /yarn-error.log
