+++ app/javascript/packs/checking.js
