+++ app/javascript/packs/progress_bar.js
