+++ package.json
@@ -5,7 +5,7 @@
     "@rails/actioncable": "^6.0.0",
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
-    "@rails/webpacker": "5.2.1",
+    "@rails/webpacker": "5.4.3",
     "bootstrap": "^5.1.3",
     "turbolinks": "^5.2.0",
     "webpack": "^4.46.0",
@@ -14,8 +14,5 @@
   "version": "0.1.0",
   "devDependencies": {
     "webpack-dev-server": "^3"
-  },
-  "engines": {
-    "node": "10.x"
   }
 }
